ansible/roles/app-server/files/app/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
prometheus-client==0.17.1